feat: Implement match list filtering
Adds support for the `where` option in match listing, filtering matches 
as requested. Supports both out-of-the-box server-side filtering where 
the query is not possible with basic Firestore indexes as well as 
enabling composite indexes. Makes use of an additional `isGameover` 
field in the metadata document to simplify look-up of matches that are 
or are not over.