feat: rename deprecated storage API methods
This renames the class methods `listGames` to `listMatches` and `createGame` to `createMatches` in
line with deprecations introduced in boardgame.io@0.41.

BREAKING CHANGE: Versions of boardgame.io <0.41.1 are no longer supported and other uses that
directly relied on calling `listGames` or `createGame` will need to be updated to use the new method
names.